Remove override from friends list & clan changes
