+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/player/chat/friend/DeleteFriend.kt
@@ -15,6 +15,6 @@ data class DeleteFriend(val friend: String) : CancellableEvent() {
     }
 }
 
-fun friendsDelete(override: Boolean = true, block: DeleteFriend.(Player) -> Unit) {
-    Events.handle("delete_friend", if (override) "*" else true, override = override, handler = block)
+fun friendsDelete(block: DeleteFriend.(Player) -> Unit) {
+    Events.handle("delete_friend", true, override = false, handler = block)
 }
